# storage/swagger/swagger.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/roles": {
            "get": {
                "summary": "角色列表",
                "operationId": "listRoles",
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                }
            },
            "post": {
                "summary": "创建角色",
                "operationId": "createRole",
                "requestBody": {
                    "description": "角色数据",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "角色名称",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "角色描述",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "description": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/roles/{id}": {
            "put": {
                "summary": "更新角色",
                "operationId": "updateRole",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "角色ID",
                        "required": true
                    }
                ],
                "requestBody": {
                    "description": "角色数据",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "角色名称",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "角色描述",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "description": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "summary": "删除角色",
                "operationId": "deleteRole",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "角色ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功"
                    }
                }
            }
        }
    }
}